requirements.txt
PyQt5==5.15.11
python-dotenv==1.1.1
requests==2.32.4
google-generativeai==0.3.1
